apps/ai-engine/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.7.0
python-dotenv==1.0.1
python-multipart==0.0.18
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
redis==5.2.1
nats-py==2.9.0
httpx==0.28.0
numpy==2.2.0
scikit-learn==1.6.0
pandas==2.2.3
sentence-transformers==3.3.1
langchain==0.3.13
langchain-community==0.3.13
langchain-openai==0.3.0
faiss-cpu==1.9.0
tiktoken==0.8.0
PyPDF2==3.0.1
python-docx==1.1.2
spacy==3.8.3
openai==1.58.0
Pillow==11.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
aiofiles==24.1.0
structlog==24.4.0
prometheus-client==0.21.1
